+++ MM-PulpFree/app/error/ErrorCodes.js
@@ -3,6 +3,10 @@ const ErrorCodes = {
   AUTH: {
     EMAIL_NOT_FOUND: 1000,
     PASSWORD_INCORRECT: 1001
+  },
+  // User ERRORS (2000 - 2999):
+  USER: {
+    NOT_FOUND: 2000
   }
 };