+++ MM-PulpFree/app/routes/index.js
@@ -3,6 +3,7 @@ const app = (module.exports = require("express")());
 //ROUTES
 app.use("/auth", require("./auth"));
 app.use("/meeting", require("./meeting"));
+app.use("/users", require("./user"));
 
 //CATCH
 app.all("*", (req, res) => {